Set to better optimized type for user_gender

+++ database.sql
@@ -16,7 +16,7 @@ CREATE TABLE users (
   user_password varchar(175) NOT NULL,
   user_bio varchar(500) NOT NULL,
   user_picture varchar(175) NOT NULL,
-  user_gender int(11) NOT NULL,
+  user_gender tinyint(2) unsigned NOT NULL,
   user_uid varchar(75) NOT NULL,
   PRIMARY KEY (user_id)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
